refactor(tools): migrate HTTP tools to directory structure
- Split monolithic http-client.ts (130 lines) into organized directory structure
- Created http/types.ts with HttpToolsConfig, schemas, and HttpResponse interface
- Split 3 tools into individual files in http/tools/:
  - http-client.ts - Full HTTP client with all methods
  - http-get.ts - Simple GET requests
  - http-post.ts - Simple POST requests
- Created http/index.ts with default exports and factory function
- Updated web/index.ts to export from new http/index.js location
- All 975 tests passing
- Build successful
- No breaking changes to public API

Follows the pattern established by Confluence, web-search, and Slack tools.
Part of Phase 3 of the tools refactoring plan.
