+++ packages/tools/src/web/index.ts
@@ -4,7 +4,7 @@
  * Tools for web interactions, scraping, and URL manipulation.
  */
 
-export * from './http-client.js';
+export * from './http/index.js';
 export * from './scraper.js';
 export * from './html-parser.js';
 export * from './url-validator.js';
